Make route directories to begin with small letter
